+++ code/packages/rust/ruby-parser/src/_grammar.rs
@@ -28,6 +28,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"return_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"break_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"next_statement"#.to_string() },
+                GrammarElement::RuleReference { name: r#"multi_assignment"#.to_string() },
                 GrammarElement::RuleReference { name: r#"modifier_statement"#.to_string() },
                 GrammarElement::RuleReference { name: r#"assignment"#.to_string() },
                 GrammarElement::RuleReference { name: r#"method_with_block"#.to_string() },
@@ -36,6 +37,25 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"expression_stmt"#.to_string() },
             ] },
             line_number: 28,
+        },
+        GrammarRule {
+            name: r#"multi_assignment"#.to_string(),
+            body: GrammarElement::Sequence { elements: vec![
+                GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                        GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
+                    ] }) },
+                GrammarElement::TokenReference { name: r#"EQUALS"#.to_string() },
+                GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                GrammarElement::Repetition { element: Box::new(GrammarElement::Sequence { elements: vec![
+                        GrammarElement::TokenReference { name: r#"COMMA"#.to_string() },
+                        GrammarElement::RuleReference { name: r#"expression"#.to_string() },
+                    ] }) },
+            ] },
+            line_number: 55,
         },
         GrammarRule {
             name: r#"modifier_statement"#.to_string(),
@@ -54,7 +74,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 64,
+            line_number: 91,
         },
         GrammarRule {
             name: r#"def_statement"#.to_string(),
@@ -72,7 +92,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 65,
+            line_number: 92,
         },
         GrammarRule {
             name: r#"class_statement"#.to_string(),
@@ -85,7 +105,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 70,
+            line_number: 97,
         },
         GrammarRule {
             name: r#"module_statement"#.to_string(),
@@ -98,7 +118,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 71,
+            line_number: 98,
         },
         GrammarRule {
             name: r#"method_with_block"#.to_string(),
@@ -120,7 +140,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"block"#.to_string() },
             ] },
-            line_number: 73,
+            line_number: 100,
         },
         GrammarRule {
             name: r#"block"#.to_string(),
@@ -128,7 +148,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::RuleReference { name: r#"do_block"#.to_string() },
                 GrammarElement::RuleReference { name: r#"brace_block"#.to_string() },
             ] },
-            line_number: 74,
+            line_number: 101,
         },
         GrammarRule {
             name: r#"do_block"#.to_string(),
@@ -141,7 +161,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 75,
+            line_number: 102,
         },
         GrammarRule {
             name: r#"brace_block"#.to_string(),
@@ -151,7 +171,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"statement"#.to_string() }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 76,
+            line_number: 103,
         },
         GrammarRule {
             name: r#"block_params"#.to_string(),
@@ -164,7 +184,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"|"#.to_string() },
             ] },
-            line_number: 77,
+            line_number: 104,
         },
         GrammarRule {
             name: r#"return_statement"#.to_string(),
@@ -172,7 +192,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"return"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 79,
+            line_number: 106,
         },
         GrammarRule {
             name: r#"break_statement"#.to_string(),
@@ -180,7 +200,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"break"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 80,
+            line_number: 107,
         },
         GrammarRule {
             name: r#"next_statement"#.to_string(),
@@ -188,7 +208,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Literal { value: r#"next"#.to_string() },
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"expression"#.to_string() }) },
             ] },
-            line_number: 81,
+            line_number: 108,
         },
         GrammarRule {
             name: r#"params"#.to_string(),
@@ -199,7 +219,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"NAME"#.to_string() },
                     ] }) },
             ] },
-            line_number: 82,
+            line_number: 109,
         },
         GrammarRule {
             name: r#"if_statement"#.to_string(),
@@ -216,7 +236,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 83,
+            line_number: 110,
         },
         GrammarRule {
             name: r#"elsif_clause"#.to_string(),
@@ -230,7 +250,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 84,
+            line_number: 111,
         },
         GrammarRule {
             name: r#"else_clause"#.to_string(),
@@ -241,7 +261,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"statement"#.to_string() },
                     ] }) },
             ] },
-            line_number: 85,
+            line_number: 112,
         },
         GrammarRule {
             name: r#"unless_statement"#.to_string(),
@@ -256,7 +276,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::Optional { element: Box::new(GrammarElement::RuleReference { name: r#"else_clause"#.to_string() }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 86,
+            line_number: 113,
         },
         GrammarRule {
             name: r#"while_statement"#.to_string(),
@@ -269,7 +289,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 87,
+            line_number: 114,
         },
         GrammarRule {
             name: r#"until_statement"#.to_string(),
@@ -282,7 +302,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Literal { value: r#"end"#.to_string() },
             ] },
-            line_number: 88,
+            line_number: 115,
         },
         GrammarRule {
             name: r#"assignment"#.to_string(),
@@ -299,7 +319,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::RuleReference { name: r#"expression"#.to_string() },
             ] },
-            line_number: 108,
+            line_number: 135,
         },
         GrammarRule {
             name: r#"method_call"#.to_string(),
@@ -319,7 +339,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"dot_call"#.to_string() }) },
             ] },
-            line_number: 119,
+            line_number: 146,
         },
         GrammarRule {
             name: r#"dot_call"#.to_string(),
@@ -341,7 +361,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"RPAREN"#.to_string() },
                     ] }) },
             ] },
-            line_number: 120,
+            line_number: 147,
         },
         GrammarRule {
             name: r#"method_call_no_paren"#.to_string(),
@@ -356,17 +376,17 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 134,
+            line_number: 161,
         },
         GrammarRule {
             name: r#"expression_stmt"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"expression"#.to_string() },
-            line_number: 135,
+            line_number: 162,
         },
         GrammarRule {
             name: r#"expression"#.to_string(),
             body: GrammarElement::RuleReference { name: r#"ternary"#.to_string() },
-            line_number: 221,
+            line_number: 248,
         },
         GrammarRule {
             name: r#"ternary"#.to_string(),
@@ -379,7 +399,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                     ] }) },
             ] },
-            line_number: 222,
+            line_number: 249,
         },
         GrammarRule {
             name: r#"range"#.to_string(),
@@ -393,7 +413,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_or"#.to_string() },
                     ] }) },
             ] },
-            line_number: 223,
+            line_number: 250,
         },
         GrammarRule {
             name: r#"logical_or"#.to_string(),
@@ -407,7 +427,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_and"#.to_string() },
                     ] }) },
             ] },
-            line_number: 224,
+            line_number: 251,
         },
         GrammarRule {
             name: r#"logical_and"#.to_string(),
@@ -421,7 +441,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"logical_not"#.to_string() },
                     ] }) },
             ] },
-            line_number: 225,
+            line_number: 252,
         },
         GrammarRule {
             name: r#"logical_not"#.to_string(),
@@ -432,7 +452,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         ] }) }) },
                 GrammarElement::RuleReference { name: r#"comparison"#.to_string() },
             ] },
-            line_number: 232,
+            line_number: 259,
         },
         GrammarRule {
             name: r#"comparison"#.to_string(),
@@ -450,7 +470,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"sum"#.to_string() },
                     ] }) },
             ] },
-            line_number: 233,
+            line_number: 260,
         },
         GrammarRule {
             name: r#"sum"#.to_string(),
@@ -464,7 +484,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"term"#.to_string() },
                     ] }) },
             ] },
-            line_number: 234,
+            line_number: 261,
         },
         GrammarRule {
             name: r#"term"#.to_string(),
@@ -478,7 +498,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::RuleReference { name: r#"factor"#.to_string() },
                     ] }) },
             ] },
-            line_number: 235,
+            line_number: 262,
         },
         GrammarRule {
             name: r#"factor"#.to_string(),
@@ -501,7 +521,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::Repetition { element: Box::new(GrammarElement::RuleReference { name: r#"dot_call"#.to_string() }) },
             ] },
-            line_number: 245,
+            line_number: 272,
         },
         GrammarRule {
             name: r#"unary_minus"#.to_string(),
@@ -509,7 +529,7 @@ pub fn parser_grammar() -> ParserGrammar {
                 GrammarElement::TokenReference { name: r#"MINUS"#.to_string() },
                 GrammarElement::RuleReference { name: r#"factor"#.to_string() },
             ] },
-            line_number: 246,
+            line_number: 273,
         },
         GrammarRule {
             name: r#"symbol_literal"#.to_string(),
@@ -521,7 +541,7 @@ pub fn parser_grammar() -> ParserGrammar {
                         GrammarElement::TokenReference { name: r#"STRING"#.to_string() },
                     ] }) },
             ] },
-            line_number: 247,
+            line_number: 274,
         },
         GrammarRule {
             name: r#"array_literal"#.to_string(),
@@ -536,7 +556,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACKET"#.to_string() },
             ] },
-            line_number: 248,
+            line_number: 275,
         },
         GrammarRule {
             name: r#"hash_literal"#.to_string(),
@@ -551,7 +571,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     ] }) },
                 GrammarElement::TokenReference { name: r#"RBRACE"#.to_string() },
             ] },
-            line_number: 249,
+            line_number: 276,
         },
         GrammarRule {
             name: r#"hash_entry"#.to_string(),
@@ -567,7 +587,7 @@ pub fn parser_grammar() -> ParserGrammar {
                     GrammarElement::RuleReference { name: r#"expression"#.to_string() },
                 ] },
             ] },
-            line_number: 250,
+            line_number: 277,
         },
     ],
         version: 1,